+++ cupy_backends/cuda/libs/nccl.pyx
@@ -11,7 +11,7 @@ from libcpp cimport vector
 from cupy_backends.cuda.api cimport driver
 from cupy_backends.cuda.api cimport runtime
 
-cdef extern from 'cupy_nccl.h':
+cdef extern from '../../cupy_nccl.h':
     ctypedef struct ncclComm:
         pass
     ctypedef ncclComm* ncclComm_t
